Update date_of_birth validation logic
