+++ app/models/api/trainee_attributes.rb
@@ -3,6 +3,8 @@
 module Api
   class TraineeAttributes
     include ActiveModel::Model
+    include ActiveModel::Attributes
+    include DateOfBirthValidatable
 
     ATTRIBUTES = %i[
       first_names
@@ -34,9 +36,18 @@ module Api
     validates(*ATTRIBUTES, presence: true)
     validates :first_names, :last_name, length: { maximum: 50 }
     validates :middle_names, length: { maximum: 50 }, allow_nil: true
-    validates :date_of_birth, date_of_birth: true
     validates :sex, inclusion: { in: Trainee.sexes.keys }
 
-    accepts_nested_attributes_for :placements, :degrees
+    def placements_attributes=(attributes)
+      @placements = attributes.map do |placement_attributes|
+        PlacementAttributes.new(placement_attributes)
+      end
+    end
+
+    def degrees_attributes=(attributes)
+      @degrees = attributes.map do |degree_attributes|
+        DegreeAttributes.new(degree_attributes)
+      end
+    end
   end
 end
